requirements.txt
aioredis==1.2.0
aniso8601==3.0.2
asgiref==3.1.4
asn1crypto==0.24.0
astroid==2.1.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
certifi==2018.8.24
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-allauth==0.37.1
django-cors-headers==2.4.0
django-import-export==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
environs==5.1.0
et-xmlfile==1.0.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
hiredis==1.0.0
httplib2==0.12.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
marshmallow==2.19.5
mccabe==0.6.1
msgpack==0.6.1
oauth2==1.9.0.post1
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
promise==2.2.1
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
soupsieve==1.6.2
sqlparse==0.3.0
tablib==0.12.1
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.23
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
